requirements.txt
pandas
numpy
pdfminer.six
docx2txt
scikit-learn
transformers
streamlit
gensim
nltk